+++ src/main/java/org/nrg/containers/model/xnat/XnatFile.java
@@ -37,6 +37,8 @@ public class XnatFile extends XnatModelObject {
         } else {
             this.uri = parentUri + "/files/" + name;
         }
+
+        this.label = name;
         this.name = name;
         this.path = path;
         this.tags = Arrays.asList(tagsCsv.split(","));